Updated API schema, added forms to all text inputs, getting started with password change
