+++ frontend/src/locales/en.json
@@ -4,6 +4,7 @@
     "error400": "Request Failure (400)",
     "error404Page": "Page Not Found",
     "invalidEmail": "Invalid email address!",
+    "noChangesMade": "You have not made any changes!",
     "noError": "No errors found at this moment",
     "noMultiFile": "Multiple files are not allowed!",
     "tooManyRequests": "Too many network requests! Please try again later!",
@@ -108,12 +109,14 @@
     "currentUsername": "Your Current Username",
     "diaglogAvatarDelete": "Are you sure you want to delete your avatar?",
     "diaglogBannerDelete": "Are you sure you want to delete your banner?",
-    "diaglogUsernameUpdate": "Are you sure you want to change your username to ",
     "diaglogEmailUpdate": "Are you sure you want to change your email to ",
+    "diaglogPasswordUpdate": "Are you sure you want to update your password?",
+    "diaglogUsernameUpdate": "Are you sure you want to change your username to ",
     "dontHaveAccount": "Don't have an account?",
     "email": "Email",
     "favoriteGame": "Favorite Game",
     "forget": "Forgot your username and/or password?",
+    "lastPasswordChange": "Last Password Change",
     "lastUsername": "Your Previous Username",
     "lastUsernameChange": "Last Username Change",
     "location": "Location",
@@ -124,7 +127,7 @@
     "newPassword": "New Password",
     "newUsername": "New Username",
     "noBanner": "Banner not found!",
-    "oldPassword": "Old Password",
+    "oldPassword": "Current Password",
     "profile": "Profile",
     "pronoun": "Pronoun",
     "pronounHe": "He / Him",
@@ -132,10 +135,12 @@
     "pronounThey": "They / Their",
     "pronounZe": "Ze / Hir",
     "password": "Password",
+    "passwordChangeNotice": "NOTICE: After you've changed your password, please note that you will need to log in again for security purposes.",
     "register": "Register",
     "remember": "Remember Me",
     "signature": "Signature",
     "signup": "Sign up!",
+    "social": "Social Media",
     "siteSettings": "Site Settings",
     "title": "Title",
     "updateDone": "Successfully updated user preferences!",
