+++ frontend/src/user/preferences/UserProfile.tsx
@@ -21,7 +21,7 @@ export default function UserProfileSettings(props: ProfileSettingsProps) {
   const { t } = useTranslation();
 
   // Context
-  const { isMobile, isSmallMobile } = useContext(GlobalContext);
+  const { isMobile, isSmallMobile, toast } = useContext(GlobalContext);
 
   // State
   const [inputs, setInputs] = useState<User>({});
@@ -33,223 +33,229 @@ export default function UserProfileSettings(props: ProfileSettingsProps) {
   // Output
   return (
     <Box mx={4} my={2} width="100%">
-      <Grid container direction="column" alignContent="stretch">
-        {/* -------------------- Title -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.title')}</b>
-          </LumaText>
-          <LumaInput
-            name="title"
-            fullWidth
-            size="small"
-            disabled={props.loading}
-            value={inputs.title ?? props.user?.title}
-            onChange={handleInputChange}
-            inputProps={{ maxLength: CF.MAX_128 }}
-          />
-          <Box width="100%" textAlign="right" mr={2} mb={1}>
+      <form onSubmit={handleFormSubmit}>
+        <Grid container direction="column" alignContent="stretch">
+          {/* -------------------- Title -------------------- */}
+          <Grid container item direction="column">
             <LumaText variant="body2">
-              {`${inputs.title?.length ?? props?.user?.title?.length ?? 0}/${
-                CF.MAX_128
-              }`}
+              <b>{t('user.title')}</b>
             </LumaText>
-          </Box>
-        </Grid>
-        {/* -------------------- Signature -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.signature')}</b>
-          </LumaText>
-          <LumaInput
-            name="signature"
-            multiline
-            fullWidth
-            rows={3}
-            size="small"
-            disabled={props.loading}
-            value={inputs.signature ?? props.user?.signature}
-            onChange={handleInputChange}
-            inputProps={{ maxLength: CF.MAX_X24 }}
-          />
-          <Box width="100%" textAlign="right" mr={2} mb={1}>
-            <LumaText variant="body2">
-              {`${
-                inputs.signature?.length ?? props?.user?.signature?.length ?? 0
-              }/${CF.MAX_X24}`}
-            </LumaText>
-          </Box>
-        </Grid>
-        {/* -------------------- Bio -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.bio')}</b>
-          </LumaText>
-          <LumaInput
-            name="bio"
-            multiline
-            fullWidth
-            rows={3}
-            size="small"
-            disabled={props.loading}
-            value={inputs.bio ?? props.user?.bio}
-            onChange={handleInputChange}
-            inputProps={{ maxLength: CF.MAX_512 }}
-          />
-          <Box width="100%" textAlign="right" mr={2} mb={1}>
-            <LumaText variant="body2">
-              {`${inputs.bio?.length ?? props?.user?.bio?.length ?? 0}/${
-                CF.MAX_512
-              }`}
-            </LumaText>
-          </Box>
-        </Grid>
-        {/* -------------------- Birthday -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.birthday')}</b>
-          </LumaText>
-          <Grid container item direction="row" alignItems="center">
             <LumaInput
-              name="birthday"
-              type="date"
-              disabled={props.loading}
-              value={inputs.birthday ?? props.user?.birthday ?? ''}
-              sx={{ width: 180 }}
-              onChange={handleInputChange}
-              onBlur={handleInputChange}
-              InputLabelProps={{ shrink: true }}
-            />
-            <LumaButton
+              name="title"
+              fullWidth
               size="small"
-              sx={{ height: '2rem' }}
-              color="secondary"
-              onClick={() => handleButton('birthday')}
-            >
-              {t('main.clear')}
-            </LumaButton>
+              disabled={props.loading}
+              value={inputs.title ?? props.user?.title}
+              onChange={handleInputChange}
+              inputProps={{ maxLength: CF.MAX_128 }}
+            />
+            <Box width="100%" textAlign="right" mr={2} mb={1}>
+              <LumaText variant="body2">
+                {`${inputs.title?.length ?? props?.user?.title?.length ?? 0}/${
+                  CF.MAX_128
+                }`}
+              </LumaText>
+            </Box>
           </Grid>
-          <Box width="100%" textAlign="right" mr={2} mb={1} height="1rem" />
-        </Grid>
-        {/* -------------------- Pronoun -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.pronoun')}</b>
-          </LumaText>
-          <LumaSelect
-            name="pronoun"
-            fullWidth
-            disabled={props.loading}
-            value={inputs.pronoun ?? props.user?.pronoun ?? 'Unspecified'}
-            onChange={handleMenu}
-          >
-            <option value="Unspecified">{t('main.unspecified')}</option>
-            <option value="He">{t('user.pronounHe')}</option>
-            <option value="She">{t('user.pronounShe')}</option>
-            <option value="They">{t('user.pronounThey')}</option>
-            <option value="Ze">{t('user.pronounZe')}</option>
-          </LumaSelect>
-          <Box width="100%" textAlign="right" mr={2} mb={1} height="1rem" />
-        </Grid>
-        {/* -------------------- Country -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.country')}</b>
-          </LumaText>
-          <LumaSelect
-            name="country"
-            fullWidth
-            disabled={props.loading}
-            value={inputs.country ?? props.user?.country ?? 'Unspecified'}
-            onChange={handleMenu}
-          >
-            <option value="Unspecified">{t('main.unspecified')}</option>
-            {countries.map((c, k) => (
-              <option key={k} value={c}>
-                {c}
-              </option>
-            ))}
-          </LumaSelect>
-          <Box width="100%" textAlign="right" mr={2} mb={1} height="1rem" />
-        </Grid>
-        {/* -------------------- Location -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.location')}</b>
-          </LumaText>
-          <LumaInput
-            name="location"
-            fullWidth
-            size="small"
-            disabled={props.loading}
-            value={inputs.location ?? props.user?.location}
-            onChange={handleInputChange}
-            inputProps={{ maxLength: CF.MAX_128 }}
-          />
-          <Box width="100%" textAlign="right" mr={2} mb={1}>
+          {/* -------------------- Signature -------------------- */}
+          <Grid container item direction="column">
             <LumaText variant="body2">
-              {`${
-                inputs.location?.length ?? props?.user?.location?.length ?? 0
-              }/${CF.MAX_128}`}
+              <b>{t('user.signature')}</b>
             </LumaText>
+            <LumaInput
+              name="signature"
+              multiline
+              fullWidth
+              rows={3}
+              size="small"
+              disabled={props.loading}
+              value={inputs.signature ?? props.user?.signature}
+              onChange={handleInputChange}
+              inputProps={{ maxLength: CF.MAX_X24 }}
+            />
+            <Box width="100%" textAlign="right" mr={2} mb={1}>
+              <LumaText variant="body2">
+                {`${
+                  inputs.signature?.length ??
+                  props?.user?.signature?.length ??
+                  0
+                }/${CF.MAX_X24}`}
+              </LumaText>
+            </Box>
+          </Grid>
+          {/* -------------------- Bio -------------------- */}
+          <Grid container item direction="column">
+            <LumaText variant="body2">
+              <b>{t('user.bio')}</b>
+            </LumaText>
+            <LumaInput
+              name="bio"
+              multiline
+              fullWidth
+              rows={3}
+              size="small"
+              disabled={props.loading}
+              value={inputs.bio ?? props.user?.bio}
+              onChange={handleInputChange}
+              inputProps={{ maxLength: CF.MAX_512 }}
+            />
+            <Box width="100%" textAlign="right" mr={2} mb={1}>
+              <LumaText variant="body2">
+                {`${inputs.bio?.length ?? props?.user?.bio?.length ?? 0}/${
+                  CF.MAX_512
+                }`}
+              </LumaText>
+            </Box>
+          </Grid>
+          {/* -------------------- Birthday -------------------- */}
+          <Grid container item direction="column">
+            <LumaText variant="body2">
+              <b>{t('user.birthday')}</b>
+            </LumaText>
+            <Grid container item direction="row" alignItems="center">
+              <LumaInput
+                name="birthday"
+                type="date"
+                disabled={props.loading}
+                value={inputs.birthday ?? props.user?.birthday ?? ''}
+                sx={{ width: 180 }}
+                onChange={handleInputChange}
+                onBlur={handleInputChange}
+                InputLabelProps={{ shrink: true }}
+              />
+              <LumaButton
+                size="small"
+                sx={{ height: '2rem' }}
+                color="secondary"
+                onClick={() => handleButton('birthday')}
+              >
+                {t('main.clear')}
+              </LumaButton>
+            </Grid>
+            <Box width="100%" textAlign="right" mr={2} mb={1} height="1rem" />
+          </Grid>
+          {/* -------------------- Pronoun -------------------- */}
+          <Grid container item direction="column">
+            <LumaText variant="body2">
+              <b>{t('user.pronoun')}</b>
+            </LumaText>
+            <LumaSelect
+              name="pronoun"
+              fullWidth
+              disabled={props.loading}
+              value={inputs.pronoun ?? props.user?.pronoun ?? 'Unspecified'}
+              onChange={handleMenu}
+            >
+              <option value="Unspecified">{t('main.unspecified')}</option>
+              <option value="He">{t('user.pronounHe')}</option>
+              <option value="She">{t('user.pronounShe')}</option>
+              <option value="They">{t('user.pronounThey')}</option>
+              <option value="Ze">{t('user.pronounZe')}</option>
+            </LumaSelect>
+            <Box width="100%" textAlign="right" mr={2} mb={1} height="1rem" />
+          </Grid>
+          {/* -------------------- Country -------------------- */}
+          <Grid container item direction="column">
+            <LumaText variant="body2">
+              <b>{t('user.country')}</b>
+            </LumaText>
+            <LumaSelect
+              name="country"
+              fullWidth
+              disabled={props.loading}
+              value={inputs.country ?? props.user?.country ?? 'Unspecified'}
+              onChange={handleMenu}
+            >
+              <option value="Unspecified">{t('main.unspecified')}</option>
+              {countries.map((c, k) => (
+                <option key={k} value={c}>
+                  {c}
+                </option>
+              ))}
+            </LumaSelect>
+            <Box width="100%" textAlign="right" mr={2} mb={1} height="1rem" />
+          </Grid>
+          {/* -------------------- Location -------------------- */}
+          <Grid container item direction="column">
+            <LumaText variant="body2">
+              <b>{t('user.location')}</b>
+            </LumaText>
+            <LumaInput
+              name="location"
+              fullWidth
+              size="small"
+              disabled={props.loading}
+              value={inputs.location ?? props.user?.location}
+              onChange={handleInputChange}
+              inputProps={{ maxLength: CF.MAX_128 }}
+            />
+            <Box width="100%" textAlign="right" mr={2} mb={1}>
+              <LumaText variant="body2">
+                {`${
+                  inputs.location?.length ?? props?.user?.location?.length ?? 0
+                }/${CF.MAX_128}`}
+              </LumaText>
+            </Box>
+          </Grid>
+          {/* -------------------- Favorite Game -------------------- */}
+          <Grid container item direction="column">
+            <LumaText variant="body2">
+              <b>{t('user.favoriteGame')}</b>
+            </LumaText>
+            <LumaInput
+              name="favorite_game"
+              fullWidth
+              size="small"
+              disabled={props.loading}
+              value={inputs.favorite_game ?? props.user?.favorite_game}
+              onChange={handleInputChange}
+              inputProps={{ maxLength: CF.MAX_064 }}
+            />
+            <Box width="100%" textAlign="right" mr={2} mb={1}>
+              <LumaText variant="body2">
+                {`${
+                  inputs.favorite_game?.length ??
+                  props?.user?.favorite_game?.length ??
+                  0
+                }/${CF.MAX_064}`}
+              </LumaText>
+            </Box>
+          </Grid>
+          {/* -------------------- Submit/reset buttons -------------------- */}
+          <Box my={2} width="100%">
+            <LumaButton
+              disabled={props.loading || noChange}
+              color="secondary"
+              variant="contained"
+              size={isMobile ? 'medium' : 'large'}
+              type="submit"
+              fullWidth={isSmallMobile}
+              sx={
+                isSmallMobile
+                  ? undefined
+                  : { px: props.loading ? 4 : undefined }
+              }
+            >
+              {props.loading ? (
+                <CircularProgress size={26} color="secondary" />
+              ) : (
+                t('main.applyChanges')
+              )}
+            </LumaButton>
+            <LumaButton
+              disabled={props.loading || noChange}
+              color="primary"
+              variant="contained"
+              size={isMobile ? 'medium' : 'large'}
+              fullWidth={isSmallMobile}
+              onClick={handleReset}
+              sx={isSmallMobile ? { mt: 2 } : { ml: 2 }}
+            >
+              {t('main.reset')}
+            </LumaButton>
           </Box>
         </Grid>
-        {/* -------------------- Favorite Game -------------------- */}
-        <Grid container item direction="column">
-          <LumaText variant="body2">
-            <b>{t('user.favoriteGame')}</b>
-          </LumaText>
-          <LumaInput
-            name="favorite_game"
-            fullWidth
-            size="small"
-            disabled={props.loading}
-            value={inputs.favorite_game ?? props.user?.favorite_game}
-            onChange={handleInputChange}
-            inputProps={{ maxLength: CF.MAX_064 }}
-          />
-          <Box width="100%" textAlign="right" mr={2} mb={1}>
-            <LumaText variant="body2">
-              {`${
-                inputs.favorite_game?.length ??
-                props?.user?.favorite_game?.length ??
-                0
-              }/${CF.MAX_064}`}
-            </LumaText>
-          </Box>
-        </Grid>
-        {/* -------------------- Submit/reset buttons -------------------- */}
-        <Box my={2} width="100%">
-          <LumaButton
-            disabled={props.loading || noChange}
-            color="secondary"
-            variant="contained"
-            size={isMobile ? 'medium' : 'large'}
-            onClick={handleSubmit}
-            fullWidth={isSmallMobile}
-            sx={
-              isSmallMobile ? undefined : { px: props.loading ? 4 : undefined }
-            }
-          >
-            {props.loading ? (
-              <CircularProgress size={26} color="secondary" />
-            ) : (
-              t('main.applyChanges')
-            )}
-          </LumaButton>
-          <LumaButton
-            disabled={props.loading || noChange}
-            color="primary"
-            variant="contained"
-            size={isMobile ? 'medium' : 'large'}
-            fullWidth={isSmallMobile}
-            onClick={handleReset}
-            sx={isSmallMobile ? { mt: 2 } : { ml: 2 }}
-          >
-            {t('main.reset')}
-          </LumaButton>
-        </Box>
-      </Grid>
+      </form>
     </Box>
   );
   // Handles
@@ -300,7 +306,16 @@ export default function UserProfileSettings(props: ProfileSettingsProps) {
     }
     setInputs({ ...inputs, ...data });
   }
-  function handleSubmit() {
+  function handleFormSubmit(e: React.FormEvent<HTMLFormElement>) {
+    e.preventDefault();
+    // Do not make changes under certain circumstances
+    if (noChange) {
+      toast(t('error.noChangesMade'), 'error');
+      return;
+    }
+    if (props.loading) {
+      return;
+    }
     // Clean up any undefined values
     const output = Object.fromEntries(
       Object.entries(inputs).filter(([_, value]) => value !== undefined)
